backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
gunicorn==21.2.0
sqlalchemy>=2.0.36
pymysql==1.1.0
cryptography==41.0.7
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic>=2.5.3
pydantic-core>=2.14.6
pydantic-settings>=2.1.0
httpx==0.26.0
apscheduler==3.10.4
python-dotenv==1.0.0
pyyaml==6.0.3
slowapi==0.1.9
